% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Data Function (no final description)}
\usage{
get_data(data, download = "Environment")
}
\arguments{
\item{data}{The data specified in the show_data function}

\item{download}{Specifies the way, how the data is downloaded. Choose from c("Environment", "Local")}
}
\value{
Returns the data from the
}
\description{
Get Data Function (no final description)
}
\examples{
show_date(tag = 'Umwelt und Klima') \%>\% get_data()
}
