% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_data.R
\name{show_data}
\alias{show_data}
\title{show data function}
\usage{
show_data(external = TRUE, tag = NULL, format = NULL, message = TRUE)
}
\arguments{
\item{external}{boolean to include/exclude external datasets (not hosted by OpenData Konstanz)}

\item{tag}{vector of strings to filter for specific tags/groups}

\item{format}{vector of strings to filter for specific data formats}

\item{message}{boolean to get description of returned data set}
}
\value{
dataframe/tibble with links and information to available data sets
}
\description{
show data function
}
